Update EBCDIC encodings to use 64-bit safe def

+++ lib/ibm1122.h
@@ -92,7 +92,7 @@ static const unsigned short ibm1122_2uni[256] =
 };
 
 static int
-ibm1122_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+ibm1122_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
    unsigned char c = *s;
    *pwc = (ucs4_t) ibm1122_2uni[c];
@@ -145,7 +145,7 @@ static const struct gap ibm1122_idx[] =
 };
 
 static int
-ibm1122_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+ibm1122_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
    const struct gap* p2gap = ibm1122_idx;
    for (; p2gap->start < 0xffff ; p2gap++) {
