+++ lib/ibm1143.h
@@ -90,7 +90,7 @@ static const unsigned short ibm1143_2uni[256] =
 };
 
 static int
-ibm1143_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+ibm1143_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c = *s;
   *pwc = (ucs4_t) ibm1143_2uni[c];
@@ -134,7 +134,7 @@ static const unsigned char ibm1143_page00[] = {
 };
 
 static int
-ibm1143_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+ibm1143_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
    unsigned char c = 0;
    if (wc >= 0x0 && wc < 0x0100)
